requirements.txt
amqp==2.5.1
appnope==0.1.0
asn1crypto==0.24.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
bcrypt==3.1.7
billiard==3.6.1.0
boto3==1.9.210
botocore==1.12.210
cachetools==3.1.1
celery==4.3.0
celery-redbeat==0.13.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.4
django-appconf==1.0.3
django-celery-email==2.0.2
django-compat==1.0.15
django-cors-headers==3.1.0
django-extensions==2.2.1
django-import-export==1.2.0
django-localflavor==2.2
django-object-actions==1.1.0
django-phonenumber-field==3.0.1
django-rest-swagger==2.2.0
django-storages==1.7.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
docutils==0.15.2
ecdsa==0.13.2
et-xmlfile==1.0.1
Fabric3==1.14.post1
future==0.17.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
grep==0.3.2
gunicorn==19.9.0
html2text==2019.8.11
httplib2==0.13.1
idna==2.8
importlib-metadata==0.19
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.4
MarkupSafe==1.1.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
phonenumberslite==8.10.17
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.3.0
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-jose==3.0.1
python-stdnum==1.11
pytz==2019.2
PyYAML==5.1.2
raven==6.10.0
redis==3.3.7
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
tenacity==5.1.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.2
